Tidy this work against our own preferences
A pass over the new code against AGENTS.md, before anyone else reads it:

The column-list code moved out of the table file into its own. That file had
grown to 955 lines, forty short of the size we fail a build at, and reading a
chosen set of columns is its own idea with its own name. Its tests moved with
it.

The helper that skips a pointless read is now named for what it checks — that
no row can match — rather than sounding like it checks whether the rows are
empty. The check behind it is no longer offered to callers, since nothing
outside used it; two tests that reached for it now watch what actually happens
instead.

The ledger's newest-first list still built its own ending — the filter, the
order and the cap — as did the activity log and the column lists. All three
now share one piece, and say what they return by its proper name rather than
spelling the shape out again.

Three long comments cut back to the one thing each was there to say.
