+++ src/shared/db/api-keys.ts
@@ -16,7 +16,8 @@ import type { BlindIndex, WrappedKey } from "#shared/crypto/sealed.ts";
 import { execute, executeUpdate, queryOne } from "#shared/db/client.ts";
 import { idAndCreatedSchema } from "#shared/db/common-schema.ts";
 import { defineIdTable } from "#shared/db/define-id-table.ts";
-import { col, defineTableProjection } from "#shared/db/table.ts";
+import { defineTableProjection } from "#shared/db/projection.ts";
+import { col } from "#shared/db/table.ts";
 import { equals } from "#shared/db/where-clauses.ts";
 import { nowIso } from "#shared/now.ts";
 import { getTouchOverride } from "#shared/test-overrides.ts";
