+++ test/shared/db/projection/select.test.ts
